+++ src/Makefile.gtest.include
@@ -42,12 +42,12 @@ zelcash_gtest_SOURCES += \
 	gtest/test_txid.cpp \
 	gtest/test_libzelcash_utils.cpp \
 	gtest/test_proofs.cpp \
-	gtest/test_paymentdisclosure.cpp \
 	gtest/test_pedersen_hash.cpp \
 	gtest/test_checkblock.cpp \
 	gtest/test_zip32.cpp
 if ENABLE_WALLET
 zelcash_gtest_SOURCES += \
+	wallet/gtest/test_paymentdisclosure.cpp \
 	wallet/gtest/test_wallet.cpp
 endif
