+++ src/wallet/asyncrpcoperation_sendmany.cpp
@@ -24,14 +24,13 @@
 #include "zelcash/IncrementalMerkleTree.hpp"
 #include "sodium.h"
 #include "miner.h"
+#include "wallet/paymentdisclosuredb.h"
 
 #include <array>
 #include <iostream>
 #include <chrono>
 #include <thread>
 #include <string>
-
-#include "paymentdisclosuredb.h"
 
 using namespace libzelcash;
