+++ src/wallet/asyncrpcoperation_sendmany.h
@@ -12,7 +12,7 @@
 #include "zelcash/JoinSplit.hpp"
 #include "zelcash/Address.hpp"
 #include "wallet.h"
-#include "paymentdisclosure.h"
+#include "wallet/paymentdisclosure.h"
 
 #include <array>
 #include <unordered_map>
