+++ src/wallet/asyncrpcoperation_shieldcoinbase.h
@@ -12,13 +12,12 @@
 #include "zelcash/JoinSplit.hpp"
 #include "zelcash/Address.hpp"
 #include "wallet.h"
+#include "wallet/paymentdisclosure.h"
 
 #include <unordered_map>
 #include <tuple>
 
 #include <univalue.h>
-
-#include "paymentdisclosure.h"
 
 // Default transaction fee if caller does not specify one.
 #define SHIELD_COINBASE_DEFAULT_MINERS_FEE   10000
