+++ src/wallet/paymentdisclosure.cpp
@@ -2,7 +2,7 @@
 // Distributed under the MIT software license, see the accompanying
 // file COPYING or http://www.opensource.org/licenses/mit-license.php.
 
-#include "paymentdisclosure.h"
+#include "wallet/paymentdisclosure.h"
 
 #include "key_io.h"
 #include "util.h"
