+++ src/wallet/paymentdisclosure.h
