+++ src/wallet/paymentdisclosuredb.cpp
@@ -2,7 +2,7 @@
 // Distributed under the MIT software license, see the accompanying
 // file COPYING or http://www.opensource.org/licenses/mit-license.php.
 
-#include "paymentdisclosuredb.h"
+#include "wallet/paymentdisclosuredb.h"
 
 #include "util.h"
 #include "dbwrapper.h"
