+++ src/wallet/paymentdisclosuredb.h
@@ -5,7 +5,7 @@
 #ifndef ZELCASH_PAYMENTDISCLOSUREDB_H
 #define ZELCASH_PAYMENTDISCLOSUREDB_H
 
-#include "paymentdisclosure.h"
+#include <wallet/paymentdisclosure.h>
 
 #include <cstdint>
 #include <string>
